requirements.txt
# 🏛️ Phần Mềm Quản Lý Đối Tượng Thi Hành Án
# Requirements file

# GUI Framework
PyQt6>=6.4.0

# Database
# sqlite3 is built-in with Python

# Data Processing
pandas>=1.5.0
openpyxl>=3.0.10
numpy>=1.21.0

# Document Processing
python-docx>=0.8.11
docxtpl>=0.16.7
Pillow>=9.0.0

# Date/Time Processing
python-dateutil>=2.8.0

# Excel Processing
xlrd>=2.0.0
xlwt>=1.3.0

# PDF Generation
reportlab>=3.6.0

# Security & Encryption
cryptography>=3.4.0
bcrypt>=4.0.0

# Data Validation
pydantic>=1.10.0

# Configuration Management
configparser>=5.3.0
pyyaml>=6.0.0

# File Compression
# zipfile is built-in with Python

# JSON Processing
# json is built-in with Python

# Regular Expressions
# re is built-in with Python

# Path Handling
# pathlib is built-in with Python

# AI/ML Libraries
scikit-learn>=1.1.0
xgboost>=1.6.0
joblib>=1.2.0

# API Development
fastapi>=0.95.0
uvicorn>=0.20.0
requests>=2.28.0

# Monitoring & Logging
prometheus-client>=0.15.0
elasticsearch>=8.0.0

# Development Tools
pytest>=7.0.0
black>=22.0.0
ruff>=0.1.0

# Code Quality
mypy>=1.0.0
flake8>=5.0.0

# Testing
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Packaging
pyinstaller>=5.0.0

# Development Dependencies (optional)
# jupyter>=1.0.0
# ipython>=8.0.0 